refactor: reorganize job local store implementation and remove obsolete jobs.py
