+++ src/syft_rds/client/local_store.py
@@ -4,12 +4,12 @@ from syft_core import Client as SyftBoxClient
 
 from syft_rds.client.local_stores.base import CRUDLocalStore
 from syft_rds.client.local_stores.dataset import DatasetLocalStore
-from syft_rds.client.local_stores.jobs import JobLocalStore
+from syft_rds.client.local_stores.job import JobLocalStore
 from syft_rds.client.local_stores.runtime import RuntimeLocalStore
 from syft_rds.client.local_stores.user_code import UserCodeLocalStore
+from syft_rds.client.local_stores.custom_function import CustomFunctionLocalStore
 from syft_rds.models.base import ItemBase, ItemBaseCreate, ItemBaseUpdate
 from syft_rds.models import Dataset, Job, Runtime, UserCode, CustomFunction
-from syft_rds.client.local_stores.custom_function import CustomFunctionLocalStore
 
 if TYPE_CHECKING:
     from syft_rds.client.rds_client import RDSClientConfig
