+++ src/syft_rds/client/local_stores/job.py
@@ -1,9 +1,10 @@
 import shutil
 from pathlib import Path
 from typing import Final, Type
+from uuid import UUID
 
 from syft_rds.client.local_stores.base import CRUDLocalStore
-from syft_rds.models.models import Job, JobCreate, JobUpdate
+from syft_rds.models import Job, JobCreate, JobUpdate
 
 
 class JobLocalStore(CRUDLocalStore[Job, JobCreate, JobUpdate]):
@@ -33,3 +34,14 @@ class JobLocalStore(CRUDLocalStore[Job, JobCreate, JobUpdate]):
                 )
 
         return syftbox_output_path
+
+    def delete_by_id(self, uid: UUID) -> bool:
+        """Delete a job by its UUID.
+
+        Args:
+            uid: UUID of the job to delete
+
+        Returns:
+            True if the job was deleted, False if not found
+        """
+        return self.store.delete(uid)
